# lib/api-zod/src/generated/types/usageEventInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vijay Transmission Cost Sheet Suite API
 * OpenAPI spec version: 0.1.0
 */
import type { UsageEventInputActivityState } from './usageEventInputActivityState';
import type { UsageEventInputEntityType } from './usageEventInputEntityType';
import type { UsageEventInputEventType } from './usageEventInputEventType';
import type { UsageEventInputMetadata } from './usageEventInputMetadata';

export interface UsageEventInput {
  eventType: UsageEventInputEventType;
  /**
     * @maxLength 120
     * @nullable
     */
  pagePath?: string | null;
  /** @nullable */
  activityState?: UsageEventInputActivityState;
  /** @nullable */
  entityType?: UsageEventInputEntityType;
  /**
     * @maxLength 120
     * @nullable
     */
  entityId?: string | null;
  /** @nullable */
  metadata?: UsageEventInputMetadata;
}