# lib/api-zod/src/generated/types/userUsageEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vijay Transmission Cost Sheet Suite API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUsageEvent {
  id: number;
  eventType: string;
  /** @nullable */
  pagePath?: string | null;
  /** @nullable */
  activityState?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  entityType?: string | null;
  /** @nullable */
  entityId?: string | null;
  occurredAt: string;
}