# lib/api-zod/src/generated/types/userUsagePage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vijay Transmission Cost Sheet Suite API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUsagePage {
  path: string;
  visits: number;
  firstVisitedAt?: string;
  lastVisitedAt?: string;
}