# lib/api-zod/src/generated/types/userUsageResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vijay Transmission Cost Sheet Suite API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUsageSummary } from './userUsageSummary';

export interface UserUsageResponse {
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  from: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  to: string;
  users: UserUsageSummary[];
}