# lib/api-zod/src/generated/types/userUsageSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vijay Transmission Cost Sheet Suite API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUsageEvent } from './userUsageEvent';
import type { UserUsagePage } from './userUsagePage';

export interface UserUsageSummary {
  userId: number;
  userName: string;
  email: string;
  sessionCount: number;
  activeSeconds: number;
  idleSeconds: number;
  /** @nullable */
  lastActiveAt?: string | null;
  pageVisitCount: number;
  uniquePageCount: number;
  quoteCount: number;
  totalCostGenerated: number;
  reportCount: number;
  pages: UserUsagePage[];
  recentEvents: UserUsageEvent[];
}